requirements.txt
oracledb>=2.0.0
pandas>=2.0.0
oci>=2.100.0
python-dotenv>=1.0.0
ipykernel